test version_or_branch for dev vs test env

+++ spec/ops/revision_spec.rb
@@ -2,21 +2,31 @@ require 'spec_helper'
 
 describe Ops::Revision do
 
+  let(:create_version_file) do
+    @version_file = File.join(@root, 'VERSION')
+    File.open(@version_file, 'w') { |f| f.write('v23.24.25^{}') }
+  end
+
   before do
     @root = File.dirname(__FILE__)
-    @version_file = File.join(@root, 'VERSION')
-    settings = double("settings", {:file_root => @root, :environment => 'test'})
-    File.open(@version_file, 'w') { |f| f.write('v23.24.25^{}') }
     headers = {}
+    settings = double("settings", {:file_root => @root, :environment => 'test'})
     @version = Ops::Revision.new(headers, settings)
   end
 
   after do
-    File.delete(@version_file)
+    File.delete(@version_file) unless @version_file.nil?
   end
 
-  it "should have version" do
+  it "test env should have version" do
+    create_version_file
     @version.version_or_branch.should eq("v23.24.25")
+  end
+
+  it "development env should have branch" do
+    dev_settings = double("settings", {:file_root => @root, :environment => 'development'})
+    dev_version = Ops::Revision.new({}, dev_settings)
+    dev_version.version_or_branch.should eq("master")
   end
 
 end
